質問取得時にIDも返すように変更
